pyproject.toml
[project]
name = "govuk-chat-evaluation-prototype-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[project.scripts]
govuk_chat_evaluation = "govuk_chat_evaluation.cli:main"

[dependency-groups]
dev = [
  "pyright>=1.1.396",
  "pytest>=8.3.5",
  "pytest-freezer>=0.4.9",
  "pytest-mock>=3.14.0",
  "ruff>=0.9.9",
]

# Pyright LSP pick up uv venv directory
[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.setuptools]
packages = ["govuk_chat_evaluation"]
